requirements.txt
kafka-python==2.2.15
python-dotenv==1.1.1
